Cell improve constructor performance