@kitaisreal 2023-07-04
Iterations 10000000 cells size 12.
Cells constructor 12% performance improvement.
Parse 26% performance improvement.

```
Before:

Serialize: 0.454000s
Cells constructor: 0.931897s
Parse: 0.325014s
Copy: 0.168819s
Move: 0.023402s

After:

Serialize: 0.457333s
Cells constructor: 0.817422s
Parse: 0.239329s
Copy: 0.162679s
Move: 0.022686s
```

Iterations 10000000 cells size 60.
Cells constructor 15% performance improvement.
Parse 25% performance improvement.

```
Before:

Serialize: 1.682229s
Cells constructor: 2.985829s
Parse: 1.471215s
Copy: 0.223400s
Move: 0.023331s

After:

Serialize: 1.655973s
Cells constructor: 2.519810s
Parse: 1.092458s
Copy: 0.224395s
Move: 0.023292s
```